app/core/security.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
# @Time       : 2025/5/21 14:50
# @Author     : fany
# @Project    : PyCharm
# @File       : security.py
# @Description:
from datetime import datetime, timedelta
from typing import Optional
import jwt
from fastapi import HTTPException, status
import os
import hashlib
import hmac

SECRET_KEY = os.getenv("SECRET_KEY", "your-secret-key-here-change-in-production")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24  # 24 hours


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """创建访问令牌"""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)

    to_encode.update({"exp": expire, "iat": datetime.utcnow()})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


def verify_token(token: str):
    """验证令牌"""
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except jwt.ExpiredSignatureError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token已过期",
            headers={"WWW-Authenticate": "Bearer"},
        )
    except jwt.JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="无效的token",
            headers={"WWW-Authenticate": "Bearer"},
        )


def encrypt_app_secret(app_secret: str) -> str:
    """加密存储 app_secret"""
    return hashlib.sha256(app_secret.encode()).hexdigest()


def verify_app_secret(plain_secret: str, encrypted_secret: str) -> bool:
    """验证 app_secret"""
    return encrypt_app_secret(plain_secret) == encrypted_secret
