app/services/bilin_api_service.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
# @Time       : 2025/5/22 15:45
# @Author     : fany
# @Project    : PyCharm
# @File       : bilin_api_service.py
# @Description:
import httpx
import asyncio
from typing import Dict, Any
import logging

logger = logging.getLogger(__name__)


class BilinAPIService:
    """比邻API服务 - 用于验证app_key、app_secret和回调地址"""

    def __init__(self):
        self.base_url = os.getenv("BILIN_API_BASE_URL", "https://api.bilin.example.com")
        self.timeout = 30

    async def verify_credentials(self, app_key: str, app_secret: str, callback_address: str) -> Dict[str, Any]:
        """
        验证比邻应用凭证和回调地址配置

        Args:
            app_key: 应用key
            app_secret: 应用密钥
            callback_address: 回调地址

        Returns:
            dict: 验证结果
        """
        try:
            async with httpx.AsyncClient(timeout=self.timeout) as client:
                # 1. 验证应用凭证
                auth_response = await self._verify_app_credentials(client, app_key, app_secret)
                if not auth_response.get("valid", False):
                    return {
                        "success": False,
                        "error": "应用key或密钥无效",
                        "error_code": "INVALID_CREDENTIALS"
                    }

                # 2. 验证回调地址配置
                callback_response = await self._verify_callback_address(client, app_key, app_secret, callback_address)
                if not callback_response.get("valid", False):
                    return {
                        "success": False,
                        "error": "回调地址未正确配置或不匹配",
                        "error_code": "INVALID_CALLBACK"
                    }

                return {
                    "success": True,
                    "app_info": auth_response.get("app_info", {}),
                    "callback_verified": True
                }

        except httpx.TimeoutException:
            logger.error("比邻API验证超时")
            return {
                "success": False,
                "error": "验证服务超时，请稍后重试",
                "error_code": "TIMEOUT"
            }
        except Exception as e:
            logger.error(f"比邻API验证异常: {str(e)}")
            return {
                "success": False,
                "error": "验证服务异常",
                "error_code": "SERVICE_ERROR"
            }

    async def _verify_app_credentials(self, client: httpx.AsyncClient, app_key: str, app_secret: str) -> Dict[str, Any]:
        """验证应用凭证"""
        # 模拟API调用 - 实际项目中替换为真实的比邻API
        # 这里用简单的逻辑模拟验证过程
        await asyncio.sleep(0.1)  # 模拟网络延迟

        # 模拟验证逻辑
        if len(app_key) >= 10 and len(app_secret) >= 20:
            return {
                "valid": True,
                "app_info": {
                    "app_name": "测试应用",
                    "app_id": app_key
                }
            }
        else:
            return {"valid": False}

    async def _verify_callback_address(self, client: httpx.AsyncClient, app_key: str, app_secret: str,
                                       callback_address: str) -> Dict[str, Any]:
        """验证回调地址配置"""
        # 模拟API调用 - 实际项目中替换为真实的比邻API
        await asyncio.sleep(0.1)  # 模拟网络延迟

        # 模拟验证逻辑：检查回调地址是否在比邻系统中正确配置
        if callback_address.startswith('https://') and 'callback' in callback_address:
            return {"valid": True}
        else:
            return {"valid": False}